second try

+++ db/migrate/20150408154021_create_sold_products.rb.rb
@@ -4,7 +4,7 @@ class CreateSoldProducts < ActiveRecord::Migration
       t.integer :purchase_id, default: false, null: false
       t.integer :gamekey_id, default: false, null: false
       t.integer :vat_id, default: false, null: false
-      t.decimal :price, default: false, null: false
+      t.float :price, default: false, null: false
 
       t.timestamps null: false
     end
